\name{EnKF4_0720-package}
\alias{EnKF4_0720-package}
\alias{EnKF4_0720}
\docType{package}
\title{
\packageTitle{EnKF4_0720}
}
\description{
\packageDescription{EnKF4_0720}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EnKF4_0720}
\packageIndices{EnKF4_0720}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EnKF4_0720}

Maintainer: \packageMaintainer{EnKF4_0720}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
